feat: a second pass of navigation.rs
